add kotlin to web module + refactoring
