+++ .gitignore
@@ -1,23 +1,16 @@
-# Compiled class file
-*.class
-
-# Log file
+.idea
+*.iml
+*.ipr
+*.iws
 *.log
-
-# BlueJ files
-*.ctxt
-
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files #
-*.jar
-*.war
-*.nar
-*.ear
-*.zip
-*.tar.gz
-*.rar
+*.log.*
+.classpath
+.project
+.settings
+/.gradle
+/*/.gradle
+/build
+/*/build
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
