+++ core/src/main/java/rs/kunpero/vacation/service/dto/AddVacationInfoRequestDto.java
@@ -1,7 +1,6 @@
 package rs.kunpero.vacation.service.dto;
 
 import lombok.Data;
-import lombok.ToString;
 import lombok.experimental.Accessors;
 
 import java.time.LocalDate;
@@ -9,7 +8,6 @@ import java.util.List;
 
 @Data
 @Accessors(chain = true)
-@ToString
 public class AddVacationInfoRequestDto {
     private String userId;
 
@@ -18,4 +16,9 @@ public class AddVacationInfoRequestDto {
     private LocalDate dateTo;
 
     private List<String> substitutionIdList;
+
+//    public AddVacationInfoRequestDto setSubstitutionIdList(List<String> substitutionIdList) {
+//        this.substitutionIdList = Collections.unmodifiableList(substitutionIdList);
+//        return this;
+//    }
 }
